address review
- Drop gratuitous Option wrappers in FastFieldsPluginWriter; serialize takes &mut self
- Replace PluginCheckedIndex newtype with Index::validate_plugins; add Index::custom_extensions
- Fail closed when merging indices whose source plugin sets differ
- Avoid one-letter closure variable names in plugin code
- Unify fieldnorm + postings into one InvertedIndexPlugin (in the index module); serialize consumes self, dropping the separate close() pass
- Index built-in writers (inverted index, fast fields, store) generically so any Document type is indexed zero-copy; custom plugins receive the document type-erased via ErasedDocument instead of a materialized TantivyDocument
