+++ src/postings/mod.rs
@@ -10,7 +10,6 @@ mod indexing_context;
 mod json_postings_writer;
 mod loaded_postings;
 mod per_field_postings_writer;
-pub(crate) mod plugin;
 mod postings;
 mod postings_writer;
 mod recorder;
@@ -24,9 +23,8 @@ pub(crate) use loaded_postings::LoadedPostings;
 pub(crate) use stacker::compute_table_memory_size;
 
 pub use self::block_segment_postings::BlockSegmentPostings;
-pub(crate) use self::indexing_context::IndexingContext;
+pub(crate) use self::indexing_context::{compute_initial_table_size, IndexingContext};
 pub(crate) use self::per_field_postings_writer::PerFieldPostingsWriter;
-pub use self::plugin::{PostingsPlugin, PostingsPluginWriter};
 pub use self::postings::Postings;
 pub(crate) use self::postings_writer::{serialize_postings, IndexingPosition, PostingsWriter};
 pub use self::segment_postings::SegmentPostings;
